// src/app/models/callUpdate.model.ts
export class CallUpdate {
        StatusId: number;
        Name: string;
        AgeId: number;
        GenderId: number;
        CityId: number;
        NationalityId: number;
        Appoiment: string;
        Note: string;
    constructor(
        StatusId: number,
        Name: string,
        AgeId: number,
        GenderId: number,
        CityId: number,
        NationalityId: number,
        Appoiment: string,
        Note: string,
        ) {}
}